feat: remove NativeTopology
