+++ test/functional/insert.test.js
@@ -436,37 +436,6 @@ describe('Insert', function () {
     }
   });
 
-  it('shouldCorrectlyInsertAndUpdateWithNoCallback', {
-    // Add a tag that our runner can trigger on
-    // in this case we are setting that node needs to be higher than 0.10.X to run
-    metadata: {
-      requires: { topology: ['single', 'replicaset', 'sharded', 'ssl', 'heap', 'wiredtiger'] }
-    },
-
-    test: function (done) {
-      var configuration = this.configuration;
-      var client = configuration.newClient(configuration.writeConcernMax(), { poolSize: 1 });
-      client.connect(function (err, client) {
-        var db = client.db(configuration.db);
-        var collection = db.collection('test_insert_and_update_no_callback');
-
-        // Insert the update
-        collection.insert({ i: 1 });
-        // Update the record
-        collection.update({ i: 1 }, { $set: { i: 2 } });
-
-        // Make sure we leave enough time for mongodb to record the data
-        setTimeout(function () {
-          // Locate document
-          collection.findOne({}, function (err, item) {
-            test.equal(2, item.i);
-            client.close(done);
-          });
-        }, 100);
-      });
-    }
-  });
-
   it('shouldInsertAndQueryTimestamp', {
     // Add a tag that our runner can trigger on
     // in this case we are setting that node needs to be higher than 0.10.X to run
